Add student class
