+++ person.rb
@@ -1,0 +1,19 @@
+class Person
+    def initialize(name = "unknown", age, parent_permission = true)
+        @id = Random.rand(1..1000)
+        @name = name
+        @age = age
+        @parent_permission = parent_permission
+    end
+
+    attr_reader :id
+    attr_accessor :name, :age
+
+    def of_age?
+        @age >= 18
+    end
+
+    def can_use_services?
+        of_age? || @parent_permission
+    end
+end
